+++ src/modules/symbols/copyrights.js
@@ -1,6 +1,5 @@
 import { base } from "../../const.js";
-import { addSpaceBeforeSymbol } from "../whitespace/spaces.js";
-import { addNbspAfterSymbol, replaceSpacesWithNbspAfterSymbol } from "../whitespace/nbsp.js";
+import { fixSpacingAroundSymbol } from "./symbol-utils.js";
 
 //
 
@@ -30,22 +29,6 @@ export function replaceCopyright(string, copyrightLetter, copyrightSign) {
 //
 
 /**
- * Consolidates spaces around copyright sign
- *
- * @param {string} string - The input string where marks will be replaced.
- * @param {string} copyrightSign - The sign of choice, either © or ℗
- * @returns {string} - The string with the consolidated spaces around the copyright sign
- */
-export function consolidateSpaces(string, copyrightSign) {
-  string = addSpaceBeforeSymbol(string, copyrightSign);
-  string = addNbspAfterSymbol(string, copyrightSign);
-  string = replaceSpacesWithNbspAfterSymbol(string, copyrightSign);
-  return string;
-}
-
-//
-
-/**
  * Fixes occurrences of copyright (©), and sound recording copyright (℗) in a given string.
  *
  * @param {string} string - The input string to be fixed.
@@ -53,8 +36,8 @@ export function consolidateSpaces(string, copyrightSign) {
  */
 export function fixCopyrights(string) {
   string = replaceCopyright(string, "c", base.copyright);
-  string = consolidateSpaces(string, base.copyright);
+  string = fixSpacingAroundSymbol(string, base.copyright);
   string = replaceCopyright(string, "p", base.soundRecordingCopyright);
-  string = consolidateSpaces(string, base.soundRecordingCopyright);
+  string = fixSpacingAroundSymbol(string, base.soundRecordingCopyright);
   return string;
 }