+++ src/modules/symbols/numero-sign.js
@@ -1,0 +1,9 @@
+import { base } from "../../const.js";
+import { fixSpacingAroundSymbol } from "./symbol-utils.js";
+
+
+export function fixNumeroSign(string) {
+  string = fixSpacingAroundSymbol(string, base.numeroSign);
+
+  return string;
+}