+++ src/modules/symbols/section-sign.js
@@ -1,15 +1,10 @@
 import { base } from "../../const.js";
-import { addSpaceBeforeSymbol } from "../whitespace/spaces.js";
-import { addNbspAfterSymbol, replaceSpacesWithNbspAfterSymbol } from "../whitespace/nbsp.js";
+import { fixSpacingAroundSymbol } from "./symbol-utils.js";
+
 
 export function fixSectionSign(string) {
-  string = addSpaceBeforeSymbol(string, base.sectionSign);
-  string = addNbspAfterSymbol(string, base.sectionSign);
-  string = replaceSpacesWithNbspAfterSymbol(string, base.sectionSign);
-
-  string = addSpaceBeforeSymbol(string, base.paragraphSign);
-  string = addNbspAfterSymbol(string, base.paragraphSign);
-  string = replaceSpacesWithNbspAfterSymbol(string, base.paragraphSign);
+  string = fixSpacingAroundSymbol(string, base.sectionSign);
+  string = fixSpacingAroundSymbol(string, base.paragraphSign);
 
   return string;
 }